Squashed commit of the following (up to f9bb6):     IndexParams: set currentVersion and maxVersion to 2 for properties     Change defaultXORMask for random ordering to DEFAULT_TOGGLE_MASK (from Kraken 2) instead of Random.nextLong() for repeatable behaviour     Factor out some properties keys (IndexParams and SplitterFormat) from literals to constants     Code style: factor out some string contants (file paths etc) into methods.  Remove dead comment.     Avoid allocation by reusing a ThreadLocal buffer in CanonicalPriorities     Add KmerTableBuilder3     Simplify structure in MinimizerConfig     CanonicalPriorities: ensure sparse minimizer sets like UHS's contain all forward orientation minimizers, or they won't work as a canonical scheme     Introduce TestGenerators.mAndKPairsBalanced to cover MinTable in more test cases.  Move "no minimizer" exception check in PosRankWindow to a more logical place.     Improve performance of BitRepresentation.bytesToString (repeated call to StringBuilder.size was costly)     Add CanonicalPriorities to Testing.minimizerPriorities generators     Remove the Signature ordering from this branch     Add support for canonicalizing, in principle, any minimizer ordering by wrapping MinimizerPriorities in CanonicalPriorities.     Remove unused type variable in ClassifierTest     Add more test cases for SlackenConfTest and DiscountConfTest     Tune SparkSession config for tests to improve test performance     Speed up Spark-based tests (SparkSessionTestWrapper) by using a single shared SparkSession object.     Add DiscountConfTest     Print an informative message when a HDFS file is written     Replace division by powers of 2 by a right shift where appropriate, for better performance. (JVM does not seem to perform this optimization automatically)     Remove the BitRepresentation quadToByte mechanism and instead use NTBitArrayBuilder for encoding as this is much simpler.     Rename SplitSegment to LocationSegment for clarity. Improve comments.     Rename some things that were referred to as "hash" to "bucket" and "rank", as appropriate, for consistency and clarity     Add NTBitArrayBuilder to simplify the construction of some bit patterns     Simplify ShiftScanner: use skipRow() instead of addRow(invalidMinimizer)     Optimizations and comments for MinimizerPriorities and ShiftScanner     Bugfix for pregrouped mode with normalization. Was crashing due to not extracting bucket ID from super-mer rank correctly.     Restore a custom version of LongArrays from fastutil, so that we can tune the radix sort algorithm. Set NO_REC_SMALL to 32.     Specialize KmerTableBuilder into KmerTableBuilder1 and KmerTableBuilder2, optimizing the addRow method for each case.     KmerTableBuilder improvements: build directly in the final arrays, instead of in ArrayBuilder. This avoids the copying and allocation when ArrayBuilder.result is obtained. Add "unsafe" methods to KmerTableBuilder that forgo bounds checking (instead, the user can explicitly indicate when a row begins and ends).     FastqTextInput efficiency improvement (avoid regex matching operations)     Efficiency improvement (reuse buffers) for CountedKmers.supermersWithTagsReadable     Add fastaPrefix option to Discount count command     Cleanups (NTBitArray and decoder)     Simplify MinSplitterProps and delete redundant MinSplitterTest     Convert some unit tests from scalatest+scalacheck style to pure scalacheck style, as well as simplify where possible.     New output format: output super-mers with k-mer counts and minimizers in FASTA or TSV format. The old super-mer output format (no counts) is dropped.     Bugfix for mean superkmer length calculation: was based on total abundance, rather than distinct k-mers     Repartition FileInputs after union, to spread out data from unevenly sized files evenly     Bugfix: Discount was no longer using frequency ordering by default. Also, move MinimizerOrdering.frequencyBySequence to AdvancedMinimizerConfiguration.     Compile for Java 17 by default     Simplification: import spark.implicits instead of spark.sqlContext.implicits throughout     Set scalacheck to 1.19     Set Scala to 2.12.21/2.13.18, and set Scapegoat to 3.3.2

+++ src/main/scala/com/jnpersson/kmers/minimizer/SampledFrequencies.scala
@@ -37,7 +37,7 @@ object SampledFrequencies {
         throw new Exception("Sampling based on these minimizer priorities is not possible: numMinimizers is undefined.")
       case Some(n) =>
         assert(n < Int.MaxValue)
-        new Array[Int](n.toInt)
+        new Array[Long](n.toInt)
     }
 
     for {
@@ -61,10 +61,10 @@ object SampledFrequencies {
    * @param counts Pairs of (minimizer rank, frequency).
    *               The minimizers should be a subset of those from the given template MinTable.
    */
-  def apply(table: MinTable, counts: Iterator[(Int, Int)]): SampledFrequencies = {
+  def apply(table: MinTable, counts: Iterator[(Int, Long)]): SampledFrequencies = {
     //Constructing the lookup array is done in a separate method here to let it be GC'ed after SampledFrequencies
     //has been constructed, reducing memory pressure.
-    val lookup = new Array[Int](Arrays.max(table.byPriority) + 1)
+    val lookup = new Array[Long](Arrays.max(table.byPriority) + 1)
     for { (k, v) <- counts } {
       //mapping motif to count
       lookup(table.byPriority(k)) = v
@@ -79,7 +79,7 @@ object SampledFrequencies {
  *              This table will be mutated and cannot be reused after passing into SampledFrequencies.
  * @param minimizerCounts Maps encoded minimizer to count.
  */
-final case class SampledFrequencies(table: MinTable, minimizerCounts: Array[Int]) {
+final case class SampledFrequencies(table: MinTable, minimizerCounts: Array[Long]) {
   private def width = table.width
   private def motifs: Array[Int] = table.byPriority
 
@@ -94,12 +94,12 @@ final case class SampledFrequencies(table: MinTable, minimizerCounts: Array[Int]
     val r = table.byPriority
     //Using the fastutils sort rather than scala array sort to avoid boxing of integers for this case
     IntArrays.parallelQuickSort(r, new IntComparator {
-      override def compare(k1: Int, k2: Int): Int = Integer.compare(minimizerCounts(k1), minimizerCounts(k2))
+      override def compare(k1: Int, k2: Int): Int = java.lang.Long.compare(minimizerCounts(k1), minimizerCounts(k2))
     })
     r
   }
 
-  def motifsAndCounts: Iterator[(Int, Int)] =
+  def motifsAndCounts: Iterator[(Int, Long)] =
     sortedMotifs.iterator.map(x => (x, minimizerCounts(x)))
 
   private def countUnseen(): Int = {
@@ -118,7 +118,7 @@ final case class SampledFrequencies(table: MinTable, minimizerCounts: Array[Int]
     val sum = Arrays.sum(minimizerCounts)
     val unseenCount = countUnseen()
 
-    def percent(x: Int) = Helpers.formatPerc(x.toDouble/sum)
+    def percent(x: Long) = Helpers.formatPerc(x.toDouble/sum)
 
     def ntString(p: Int) = NTBitArray.fromLong(p, width).toString
